requirements.txt
fastapi==0.118.0
fastapi[standard]
pandas
unvicorn
asyncpg
python-dotenv
tortoise-orm[asyncpg]
passlib